+++ CalPal/CalPal/Views/LogView.swift
@@ -98,7 +98,7 @@ struct LogView: View {
         var todayCalories: Int32 = 0
         for item in food {
             if Calendar.current.isDateInToday(item.date!){
-                todayCalories += item.calories
+                todayCalories += Int32(item.calories)
             }
         }
         return todayCalories
